Centeralise custom blocks for spawner and furnace minecarts
