+++ connector/src/main/java/org/geysermc/connector/entity/MinecartEntity.java
@@ -55,7 +55,7 @@ public class MinecartEntity extends Entity {
             metadata.put(EntityData.HURT_TIME, Math.min((int) (float) entityMetadata.getValue(), 15));
         }
 
-        if (!(this instanceof FurnaceMinecartEntity) && !(this instanceof SpawnerMinecartEntity)) { // Handled in the respective classes
+        if (!(this instanceof DefaultBlockMinecartEntity)) { // Handled in the DefaultBlockMinecartEntity class
             // Custom block
             if (entityMetadata.getId() == 10) {
                 metadata.put(EntityData.DISPLAY_ITEM, BlockTranslator.getBedrockBlockId((int) entityMetadata.getValue()));
